requested changes

+++ xlive/Blam/Engine/effects/effects.cpp
@@ -23,7 +23,7 @@ s_data_array* get_effects_location_table()
     return *Memory::GetAddress<s_data_array**>(0x4CE880, 0x4F506C);
 }
 
-effect_location_datum* __cdecl effect_location_get_next_valid_index(effect_datum* effect_datum, int32* out_index, int a3)
+effect_location_datum* __cdecl effect_location_get_next_valid_index(effect_datum* effect_datum, int32* out_index, int16 a3)
 {
     return INVOKE(0xA68DD, 0x9895D, effect_location_get_next_valid_index, effect_datum, out_index, a3);
 }
